+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Allow case arm statements to be formatted inline.
+- Allow simple conditionally compiled code to be formatted inline.
 
 ## [0.4.0-rc1] - 2025-02-20