+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.4.0-rc2] - 2025-03-12
+
 ### Fixed
 
 - Fixed spacing around generics with composite type parameters (e.g. `array of`, `set of`, `string[10]`).